refactor(feed): address deferred review findings on #616
Five follow-ups from the PR #616 review — all non-blocking when flagged,
bundled here because they touch overlapping surfaces (monitor loop,
claude-processor, ollama-client).

**Layer hygiene**
- claude-processor no longer takes a `sessionManager` injection.
  Replaced with an `onSummary(uuid, summary)` callback wired by
  server.js; the host decides whether to mirror summaries onto
  meta.claude.summary. The processor stays layer-pure.
- Moved `reconcileClaudeEnrichment` out of the generic
  `session-child-counter.js` into `claude-session-discovery.js`
  alongside `discoverClaudeSession`. The monitor loop now calls
  Claude logic as a black box; no Claude-specific code lives in the
  harness-agnostic monitor module.

**Correctness / scaling**
- Pane monitor tick now fans out per-session instead of awaiting
  each session serially. A slow tmux / pgrep / ps call on one pane
  can no longer stall every other session's update. Per-session
  in-flight guard absorbs the case where one session's tick runs
  longer than the interval.
- ollama-client cancels the stream reader in a `finally` so an
  aborted fetch (timeout or caller abort) releases the connection
  instead of leaving the underlying ReadableStream pinned to the
  socket. Previously only the fetch promise rejected; the reader
  and body held the connection until GC.

**Dead code**
- Dropped the `reconcileClaudePresence` back-compat alias (no external
  callers — the rename was internal to this branch).
- Dropped the unused `_presenceChanged` argument from
  `reconcileClaudeEnrichment`; it was reserved for future diff logic
  that never materialized.

**Tests**
- Moved the `reconcileClaudeEnrichment` test suite from
  `test/session-child-counter-claude.test.js` to
  `test/claude-session-discovery.test.js` to match the new home.
- Updated call sites to drop the removed argument.
