+++ lib/claude-processor.js
@@ -60,9 +60,12 @@ const SUMMARY_CONTEXT_LIMIT = 20;
  * @param {function} [opts.callOllama] - async (userPrompt, { systemPrompt }) => string.
  *                                       Optional — when omitted or when a call throws,
  *                                       summary generation is skipped silently.
- * @param {object} [opts.sessionManager] - When provided, summaries also land on the
- *                                         matching session's meta.claude.summary so
- *                                         the terminal tab tooltip can pick them up.
+ * @param {function} [opts.onSummary]  - `(uuid, summary) => void` callback fired
+ *                                       after each successful summary. The processor
+ *                                       stays layer-pure; the caller (server.js)
+ *                                       decides whether to mirror the summary onto
+ *                                       a session's meta.claude.summary. Swallowed
+ *                                       if the callback throws.
  * @param {number} [opts.pollIntervalMs=2000]
  * @param {number} [opts.idlePollIntervalMs=15000] - poll cadence when no
  *        subscribers are attached. Workers keep running at this slower
@@ -74,7 +77,7 @@ export function createClaudeProcessor({
   watchlist,
   topicBroker,
   callOllama = null,
-  sessionManager = null,
+  onSummary = null,
   pollIntervalMs = DEFAULT_POLL_INTERVAL_MS,
   idlePollIntervalMs = DEFAULT_IDLE_POLL_INTERVAL_MS,
   sliceLimit = DEFAULT_SLICE_LIMIT,
@@ -289,20 +292,15 @@ export function createClaudeProcessor({
           log.warn("claude-processor: setMeta summary failed", { uuid: w.uuid, error: err.message });
         }
 
-        // Mirror onto the matching session's meta.claude.summary so
-        // the terminal tab tooltip can read it through the normal
-        // session-updated broadcast channel.
-        if (sessionManager) {
+        // Let the host wire the summary into anywhere outside the
+        // processor's layer — e.g. mirroring it onto the matching
+        // tmux session's meta.claude.summary so the tab tooltip can
+        // pick it up. The processor itself stays layer-pure.
+        if (typeof onSummary === "function") {
           try {
-            const sessions = sessionManager.listSessions?.().sessions || [];
-            const target = sessions.find((s) => s.meta?.claude?.uuid === w.uuid);
-            if (target) {
-              const live = sessionManager.getSession?.(target.name);
-              const current = live?.meta?.claude || target.meta?.claude || {};
-              live?.setMeta?.("claude", { ...current, summary: result });
-            }
+            onSummary(w.uuid, result);
           } catch (err) {
-            log.warn("claude-processor: session meta stamp failed", { uuid: w.uuid, error: err.message });
+            log.warn("claude-processor: onSummary callback failed", { uuid: w.uuid, error: err.message });
           }
         }
       } catch (err) {
