+++ lib/claude-session-discovery.js
@@ -30,6 +30,8 @@ import { readFile } from "node:fs/promises";
 import { homedir } from "node:os";
 import { join } from "node:path";
 import { UUID_RE } from "./claude-event-transform.js";
+import { detectAgent } from "./agent-presence.js";
+import { log } from "./log.js";
 
 const SESSIONS_DIR = join(homedir(), ".claude", "sessions");
 
@@ -128,4 +130,63 @@ export async function discoverClaudeSession(panePid, deps = {}) {
     };
   }
   return null;
+}
+
+/**
+ * Populate `meta.claude.{uuid, startedAt}` from Claude's per-pid
+ * session file when Claude is running in the pane.
+ *
+ * Filesystem discovery removes the hook-install dependency: as long
+ * as the user is running a current Claude Code build, `~/.claude/
+ * sessions/<pid>.json` exists and carries the authoritative uuid.
+ * That's what the frontend feed tile reads to open the transcript,
+ * and the sparkle click resolves to a real feed instead of a picker.
+ *
+ * Invariants:
+ *  - Only runs when `detectAgent(currentCommand) === "claude"`.
+ *  - Probes every tick Claude is running. Cost is one `ps -ax` + one
+ *    `readFile` on the matching pid's session json — negligible next
+ *    to the existing `tmux list-panes` + `pgrep` the monitor already
+ *    does. The periodic probe is what lets the server self-heal a
+ *    stale persisted uuid (e.g. after a restart where a previous
+ *    build's discovery wrote the wrong uuid).
+ *  - Never clears `meta.claude` on exit — preserving the uuid after
+ *    Claude quits is what lets a watched feed tile keep resolving
+ *    the on-disk transcript. Stale uuids are replaced only when
+ *    discovery returns a new value, i.e. while Claude is alive.
+ *
+ * Lives here (in the Claude module) rather than in the generic
+ * session-child-counter so the generic monitor loop doesn't have to
+ * know about Claude-specific enrichment. The monitor calls this via
+ * the generic `reconcileAgentPresence` result.
+ *
+ * @param {object} session  katulong Session
+ * @param {string | null} currentCommand  tmux pane_current_command
+ * @param {number | null} panePid  pid of the pane's top-level process
+ * @param {{ discover?: typeof discoverClaudeSession }} [deps]
+ * @returns {Promise<boolean>}  true iff meta.claude was written
+ */
+export async function reconcileClaudeEnrichment(session, currentCommand, panePid, deps = {}) {
+  if (detectAgent(currentCommand) !== "claude") return false;
+  if (!panePid) return false;
+
+  const discover = deps.discover || discoverClaudeSession;
+  const found = await discover(panePid);
+  if (!found) return false;
+
+  const existing = session.meta?.claude?.uuid ?? null;
+  if (existing === found.uuid) return false;
+
+  try {
+    session.setMeta("claude", {
+      uuid: found.uuid,
+      startedAt: found.startedAt,
+    });
+    return true;
+  } catch (err) {
+    log.warn("Failed to write discovered meta.claude", {
+      session: session.name, error: err.message,
+    });
+    return false;
+  }
 }
