+++ lib/ollama-client.js
@@ -63,6 +63,12 @@ export function createOllamaClient({
     // Ollama streams NDJSON: one JSON object per newline, each with an
     // incremental `message.content` fragment, terminated by a final
     // `{"done": true, ...}` object. Aggregate the content fragments.
+    //
+    // We always cancel the reader in a finally — on abort (timeout /
+    // caller-initiated), reader.read() rejects but the underlying
+    // ReadableStream and its fetch body are not released until someone
+    // explicitly cancels or fully drains. A cancel here detaches the
+    // stream from the connection so undici can close the socket.
     const reader = res.body.getReader();
     const decoder = new TextDecoder();
     let buf = "";
@@ -77,18 +83,25 @@ export function createOllamaClient({
       if (typeof piece === "string") content += piece;
     };
 
-    for (;;) {
-      const { value, done } = await reader.read();
-      if (done) break;
-      buf += decoder.decode(value, { stream: true });
-      let nl;
-      while ((nl = buf.indexOf("\n")) >= 0) {
-        consumeLine(buf.slice(0, nl));
-        buf = buf.slice(nl + 1);
+    let drained = false;
+    try {
+      for (;;) {
+        const { value, done } = await reader.read();
+        if (done) { drained = true; break; }
+        buf += decoder.decode(value, { stream: true });
+        let nl;
+        while ((nl = buf.indexOf("\n")) >= 0) {
+          consumeLine(buf.slice(0, nl));
+          buf = buf.slice(nl + 1);
+        }
+      }
+      buf += decoder.decode();
+      if (buf) consumeLine(buf);
+    } finally {
+      if (!drained) {
+        try { await reader.cancel(); } catch { /* already closed */ }
       }
     }
-    buf += decoder.decode();
-    if (buf) consumeLine(buf);
 
     if (typeof content !== "string" || content.length === 0) {
       throw new Error("Ollama response missing message.content");
