+++ lib/session-child-counter.js
@@ -15,17 +15,18 @@
  * pgrep walk or new process spawn. The same `list-panes` call we already
  * make for the pane pid carries this for free.
  *
- * This module owns `session.meta.agent` (presence). Hook-driven
- * enrichment like `session.meta.claude.uuid` lives under a separate
- * per-harness namespace written by the hook route — this monitor
- * never touches it.
+ * This module owns `session.meta.agent` (presence) and is deliberately
+ * harness-agnostic. Claude-specific enrichment (e.g. writing
+ * `session.meta.claude.uuid` from `~/.claude/sessions/<pid>.json`)
+ * lives in `lib/claude-session-discovery.js`; the monitor calls into
+ * it as a black box — no Claude logic lives in this file.
  */
 
 import { execFile } from "node:child_process";
 import { log } from "./log.js";
 import { tmuxSocketArgs } from "./tmux.js";
 import { detectAgent, isClaudeCommand } from "./agent-presence.js";
-import { discoverClaudeSession } from "./claude-session-discovery.js";
+import { reconcileClaudeEnrichment } from "./claude-session-discovery.js";
 
 const DEFAULT_INTERVAL_MS = 5000;
 
@@ -149,70 +150,6 @@ export function reconcileAgentPresence(session, currentCommand) {
 }
 
 /**
- * Deprecated alias — preserves the old name for any caller that
- * imports it directly. Prefer `reconcileAgentPresence`. Safe to
- * remove once external consumers migrate.
- */
-export const reconcileClaudePresence = reconcileAgentPresence;
-
-/**
- * Populate `meta.claude.{uuid, startedAt}` from Claude's per-pid
- * session file when Claude is running in the pane.
- *
- * Filesystem discovery removes the hook-install dependency: as long
- * as the user is running a current Claude Code build, `~/.claude/
- * sessions/<pid>.json` exists and carries the authoritative uuid.
- * That's what the frontend feed tile reads to open the transcript,
- * and the sparkle click resolves to a real feed instead of a picker.
- *
- * Invariants:
- *  - Only runs when `detectAgent(currentCommand) === "claude"`.
- *  - Probes every tick Claude is running. Cost is one `ps -ax` + one
- *    `readFile` on the matching pid's session json — negligible next
- *    to the existing `tmux list-panes` + `pgrep` the monitor already
- *    does. The periodic probe is what lets the server self-heal a
- *    stale persisted uuid (e.g. after a restart where a previous
- *    build's discovery wrote the wrong uuid).
- *  - Never clears `meta.claude` on exit — preserving the uuid after
- *    Claude quits is what lets a watched feed tile keep resolving
- *    the on-disk transcript. Stale uuids are replaced only when
- *    discovery returns a new value, i.e. while Claude is alive.
- *
- * @param {object} session  katulong Session
- * @param {string | null} currentCommand  tmux pane_current_command
- * @param {number | null} panePid  pid of the pane's top-level process
- * @param {boolean} _presenceChanged  reserved for future diff logic
- * @param {{ discover?: typeof discoverClaudeSession }} [deps]
- * @returns {Promise<boolean>}  true iff meta.claude was written
- */
-export async function reconcileClaudeEnrichment(
-  session, currentCommand, panePid, _presenceChanged, deps = {}
-) {
-  if (detectAgent(currentCommand) !== "claude") return false;
-  if (!panePid) return false;
-
-  const discover = deps.discover || discoverClaudeSession;
-  const found = await discover(panePid);
-  if (!found) return false;
-
-  const existing = session.meta?.claude?.uuid ?? null;
-  if (existing === found.uuid) return false;
-
-  try {
-    session.setMeta("claude", {
-      uuid: found.uuid,
-      startedAt: found.startedAt,
-    });
-    return true;
-  } catch (err) {
-    log.warn("Failed to write discovered meta.claude", {
-      session: session.name, error: err.message,
-    });
-    return false;
-  }
-}
-
-/**
  * Start a periodic monitor that:
  *  1. Reaps dead, clientless sessions from the map.
  *  2. Inspects alive sessions' panes for child count + foreground command.
@@ -228,25 +165,48 @@ export async function reconcileClaudeEnrichment(
  * @returns {{ stop: () => void }}
  */
 export function startChildCountMonitor({ sessions, tracker, bridge, intervalMs = DEFAULT_INTERVAL_MS }) {
-  const timer = setInterval(async () => {
-    for (const [name, session] of [...sessions]) {
-      if (!session.alive) {
-        if (!tracker.hasClients(name)) {
-          sessions.delete(name);
-          log.info("Reaped dead session", { session: name });
-        }
-        continue;
+  // Per-session in-flight guard. Prevents a slow tmux/pgrep/ps on one
+  // session from piling up overlapping ticks — if the previous tick for
+  // a session hasn't settled, we skip this tick for that session only
+  // and let the others proceed.
+  const inFlight = new Set();
+
+  async function tickSession(name, session) {
+    if (!session.alive) {
+      if (!tracker.hasClients(name)) {
+        sessions.delete(name);
+        log.info("Reaped dead session", { session: name });
       }
-      const { childCount, currentCommand, panePid } = await inspectTmuxPane(session.tmuxName);
-      session.updateChildCount(childCount);
-      bridge.relay({ type: "child-count-update", session: name, count: childCount });
-      // Agent presence flips trigger a session.setMeta call, which itself
-      // fires onChange → session-updated broadcast; nothing to relay here.
-      const presenceChanged = reconcileAgentPresence(session, currentCommand);
-      // When Claude is running, try to enrich meta.claude from the
-      // per-pid session file Claude writes to disk. Hook-free so the
-      // sparkle tile opens a feed without an installer step.
-      await reconcileClaudeEnrichment(session, currentCommand, panePid, presenceChanged);
+      return;
+    }
+    const { childCount, currentCommand, panePid } = await inspectTmuxPane(session.tmuxName);
+    session.updateChildCount(childCount);
+    bridge.relay({ type: "child-count-update", session: name, count: childCount });
+    // Agent presence flips trigger a session.setMeta call, which itself
+    // fires onChange → session-updated broadcast; nothing to relay here.
+    reconcileAgentPresence(session, currentCommand);
+    // When Claude is running, try to enrich meta.claude from the
+    // per-pid session file Claude writes to disk. Hook-free so the
+    // sparkle tile opens a feed without an installer step.
+    await reconcileClaudeEnrichment(session, currentCommand, panePid);
+  }
+
+  const timer = setInterval(() => {
+    // Fan out per-session. Each session ticks independently — a slow
+    // tmux call on one pane can't stall every other session's update.
+    // The interval itself is what bounds concurrency (≥1 interval
+    // between ticks for any given session), and the inFlight guard
+    // absorbs the tail case where one session runs longer than the
+    // interval.
+    for (const [name, session] of [...sessions]) {
+      if (inFlight.has(name)) continue;
+      inFlight.add(name);
+      Promise.resolve()
+        .then(() => tickSession(name, session))
+        .catch((err) => log.warn("child-count monitor tick failed", {
+          session: name, error: err.message,
+        }))
+        .finally(() => inFlight.delete(name));
     }
   }, intervalMs);
   if (timer.unref) timer.unref();
